Week 5-6: AG-UI server, supply chain layer, mocks, hardware optimization
AG-UI server + frontend
- server.py: FastAPI SSE backend implementing AG-UI protocol (classify/retrieve/generate)
- frontend/index.html: streaming chat UI with product cards, step chips, right panel
- Fixed 4 JS bugs (renderProducts, createAssistantBubble destructuring, stepsEl ref)

Supply chain layer
- rufus/inventory.py: SQLite schema + query helpers (32K SKUs, 94K demand history rows)
- rufus/demand.py: ForecastResult/ReorderAlert dataclasses, bulk_reorder_check
- rufus/sc_rag.py: LLM context formatters for SC data
- 5 new SC intents: check_stock, reorder_alert, demand_forecast, supplier_query, sc_analytics
- graph.py: node_supply_chain + node_sc_generate, new routing branch

Mock features
- rufus/personalization.py: user profiles + preference-based reranking bias
- rufus/cart.py: session cart persisted via MemorySaver
- rufus/seller_qa.py: category Q&A templates
- New intents: gift_search, add_to_cart, view_cart (12 intents total)

Hardware optimization (RTX 5090, 32 GB VRAM)
- rufus/hardware.py: TF32 matmul + cuDNN benchmark applied globally
- BGE-M3 + CLIP: fp16 inference, batch_size=64/128
- Cross-encoder: fp16, batch_size=128 (was 32)
- LLM: think=False as direct kwarg (not in options), temperature=0, keep_alive=60m

Critical LLM fix
- qwen3:1.7b with think=False in options dict routes ALL output to thinking field
- Fix: pass as direct kwarg to ollama.chat() -> 0.3s TTFT, clean responses
- Switched default model from qwen3.5:latest to qwen3:1.7b

Data pipeline
- scripts/download_all_data.py: comprehensive multi-source downloader
- scripts/ingest_olist.py + ingest_scms.py: supply chain DB population
- scripts/ingest_amazon_reviews_full.py: all-category metadata + reviews
- scripts/train_demand_forecast.py: M5 + Olist -> 15K forecast rows

+++ rufus/clip_retriever.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 
 import torch
 
+import rufus.hardware  # apply TF32 + cuDNN benchmark globally
 from rufus.retriever import Product
 
 CLIP_MODEL_NAME = "openai/clip-vit-large-patch14"
@@ -39,7 +40,9 @@ class CLIPRetriever:
     def _load_model(self) -> None:
         from transformers import CLIPModel, CLIPProcessor
         self._processor = CLIPProcessor.from_pretrained(self.model_name)
-        self._model = CLIPModel.from_pretrained(self.model_name).to(self._device)
+        self._model = CLIPModel.from_pretrained(
+            self.model_name, torch_dtype=torch.float16   # fp16 for ViT-L/14 on RTX 5090
+        ).to(self._device)
         self._model.eval()
 
     def encode_text(self, text: str) -> list[float]:
